+++ modules/tinymce/src/core/main/ts/content/GetContentImpl.ts
@@ -22,9 +22,15 @@ const getContentFromBody = (editor: Editor, args: GetContentArgs, body: HTMLElem
   if (args.format === 'raw') {
     content = Tools.trim(TrimHtml.trimExternal(editor.serializer, body.innerHTML));
   } else if (args.format === 'text') {
-    // return empty string for text format when editor is empty to avoid bogus elements being returned in content
-    content = Zwsp.trim(body.innerText);
-    content = content === '\n' ? '' : content;
+    content = getContentFromBody(editor, { ...args, format: 'html' }, body);
+
+    if (content === '' || !Type.isString(content)) {
+      return content;
+    }
+
+    const parseBody = document.createElement('body');
+    parseBody.innerHTML = trimEmptyContents(editor, content);
+    content = Zwsp.trim(parseBody.innerText.replace(/\xa0/g, ''));
   } else if (args.format === 'tree') {
     content = editor.serializer.serialize(body, args);
   } else {